feat: Add generic SQLite repository implementation
In order to use this class to retrieve data from the database, it has
to be instantiated with the type argument T, where T inherits from
AbstractTable.

Notable changes:
- Added AbstractTable class, which satisfies the type requirements
imposed by the SQLite library. All DBObjects must inherit from this
class.
- Renamed CameraRepository to SQLiteRepository, since CameraRepository
contained the original database implementation.
- The GetAll() method now returns a IQueryable<T> instead of List<T>,
since the TableQuery doesn't impose abstract type restriction on its
type argument, whereas the Table<T> method did.
- Since the Get(int Id) method needs a concrete string with the table
name in order to do it's job, this table name is extracted from T's
TableAttribute in the SQLiteRepository constructor.
- The SQLiteRepository (or CameraRepository previously) in Camera.cs
was created twice, once in the private constructor and once during
static initialization.
