+++ CSS Server/Models/Database/DBObjects/AbstractTable.cs
@@ -1,0 +1,10 @@
+using SQLite;
+
+namespace CSS_Server.Models.Database.DBObjects
+{
+    public abstract class AbstractTable
+    {
+        [PrimaryKey, AutoIncrement]
+        public int Id { get; set; }
+    }
+}
