+++ CSS Server/Models/Database/Repositories/IRepository.cs
@@ -1,14 +1,14 @@
-﻿using System.Collections.Generic;
+using CSS_Server.Models.Database.DBObjects;
+using System.Linq;
 
 namespace CSS_Server.Models.Database
 {
-    public interface IRepository<T>
+    public interface IRepository<T> where T : AbstractTable
     {
         T Get(int id);
         void Delete(int id);
-        List<T> GetAll();
+        IQueryable<T> GetAll();
         void Update(T entity);
         void Insert(T entity);
-
     }
 }
